# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminPresentationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminPresentationRow"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AdminPresentationListResponse",
        "type": "object"
      },
      "AdminPresentationRow": {
        "properties": {
          "current_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_email": {
            "title": "Owner Email",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version_count": {
            "title": "Version Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "owner_id",
          "owner_email",
          "current_version_id",
          "version_count",
          "updated_at"
        ],
        "title": "AdminPresentationRow",
        "type": "object"
      },
...
      "EmbedResponse": {
        "properties": {
          "iframe_src": {
            "title": "Iframe Src",
            "type": "string"
          },
          "slide_count": {
            "title": "Slide Count",
            "type": "integer"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "iframe_src",
          "version_id",
          "slide_count"
        ],
        "title": "EmbedResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PresentationAccess": {
        "enum": [
          "admin",
          "owner",
          "editor",
          "user"
        ],
        "title": "PresentationAccess",
        "type": "string"
      },
      "PresentationCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "PresentationCreate",
        "type": "object"
      },
      "PresentationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PresentationRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PresentationListResponse",
        "type": "object"
      },
...
      "PresentationRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_user_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentationAccess"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VersionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "title",
          "description",
          "current_version_id",
          "created_at",
          "updated_at"
        ],
        "title": "PresentationRead",
        "type": "object"
      },
      "PresentationUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "PresentationUpdate",
        "type": "object"
      },
      "SlideRead": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "selector": {
            "title": "Selector",
            "type": "string"
          },
          "slide_index": {
            "title": "Slide Index",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "slide_index",
          "selector",
          "title"
        ],
        "title": "SlideRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entry_path": {
            "title": "Entry Path",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "presentation_id": {
            "format": "uuid",
            "title": "Presentation Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "slides": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SlideRead"
            },
            "title": "Slides",
            "type": "array"
          },
          "storage_kind": {
            "title": "Storage Kind",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "presentation_id",
          "version_number",
          "origin",
          "storage_kind",
          "entry_path",
          "sha256",
          "size_bytes",
          "created_at"
        ],
        "title": "VersionRead",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Internal presentation collaboration API (v1 scaffold).",
    "title": "promptDeck API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/presentations": {
      "get": {
        "operationId": "list_presentations_admin_api_v1_admin_presentations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPresentationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Presentations Admin",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/presentations": {
      "get": {
        "operationId": "list_presentations_api_v1_presentations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Presentations",
        "tags": [
          "presentations"
        ]
      },
      "post": {
        "operationId": "create_presentation_api_v1_presentations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Presentation",
        "tags": [
          "presentations"
        ]
      }
    },
    "/api/v1/presentations/{presentation_id}": {
      "delete": {
        "operationId": "delete_presentation_api_v1_presentations__presentation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "presentation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Presentation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Presentation",
        "tags": [
          "presentations"
        ]
      },
      "get": {
        "operationId": "get_presentation_detail_api_v1_presentations__presentation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "presentation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Presentation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Presentation Detail",
        "tags": [
          "presentations"
        ]
      },
      "patch": {
        "operationId": "update_presentation_api_v1_presentations__presentation_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "presentation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Presentation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Presentation",
        "tags": [
          "presentations"
        ]
      }
    },
    "/api/v1/presentations/{presentation_id}/embed": {
      "get": {
        "operationId": "embed_iframe_api_v1_presentations__presentation_id__embed_get",
        "parameters": [
          {
            "in": "path",
            "name": "presentation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Presentation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Embed Iframe",
        "tags": [
          "presentations"
        ]
      }
    },
...
  }
}